General & needed improvements

+++ generator.py
@@ -35,7 +35,7 @@ def generate_book_html(book_json_or_file):
 
   book_html = book_html.replace('{__chapters__}', "\n".join(chapters_html))
   book_html = book_html.replace('<p>&nbsp;</p>', '')
-  
+
   # finally, export the finished book html
   if not os.path.exists(filepath):
     os.makedirs(filepath)
@@ -113,12 +113,16 @@ def generate_book_pdf(book_json_or_file):
   os.system(pdf_command)
   return pdf_file
 
-def combine_audio(book_json, files):
+def combine_audio(book_json, files, keep_blinks=False):
   if not is_installed("ffmpeg"):
     print(f"[!] ffmpeg needs to be installed and added to PATH to combine audio files")
     return
 
-  print(f"[.] Combining audio files for {book_json['slug']}")
+  if keep_blinks:
+    action_text = "keeping originals of"
+  else:
+    action_text = "deleting"
+  print(f"[.] Combining, and {action_text}, audio files for {book_json['slug']}")
   filepath = get_book_pretty_filepath(book_json)
   filename = get_book_pretty_filename(book_json, ".m4a")
 
@@ -150,6 +154,7 @@ def combine_audio(book_json, files):
     os.remove(files_list)
   if (os.path.exists(combined_audio_file)):
     os.remove(combined_audio_file)
-  for file in files:
-    if (os.path.exists(file)):
-      os.remove(os.path.abspath(file))
+  if not (keep_blinks):
+    for file in files:
+      if (os.path.exists(file)):
+        os.remove(os.path.abspath(file))
